Summary add and search
Added summary add and search. Noticed something wrong with atoi, which
converts an emtpy string to 0. Thus, search will always work for finding
a 0 (getline retrieves an empty string at the end of the file read), but
may not return a proper result.

This has been fixed in Invoice and Summary by adding a break for empty
strings in the while loops of search and modify.