Done Invoice Item Isolated Tests and other stuff
-Completed Invoice Item Isolated Tests Add,Modify,Search

-Added pragma to disable 4290 error to make it look cleaner. This error
comes up because VS doesn't really implement  function declations that
throw specific exceptions, although it does accept it. Don't worry. This
won't do anything but make building look cleaner.

-Noticed a problem with the way we were using delimiters and breaking up
substrings. Before, this worked only for rows with a single digit
primary key, which is what our tests were based on so we didnt catch it.
This was my fault because of the way I worked with the category search
function. Mike, I have changed this in product and category modify and
search. Others, take not if you were copying this.

For substrings in between two '|'s:
It should be string.substring(lastdelim+1, nextdelim - lastdelim - 1)

-Little changes in commenting in TableInterface and other things